+++ ssssg/__init__.py
@@ -1,3 +1,3 @@
 from .config import options
 from .exceptions import SSSSGException
-from .page import IndexHandler, PageHandler
+from .page import ErrorHandler, IndexHandler, PageHandler