+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-## 0.2.0 -- 10/22/2026
+## 0.3.0 -- 10/23/2016
+
+**Added**
+
+    * Default handler for 404 and 500 pages
+    * Simple help text to get the server running and indexing
+    * default_error_title option
+
+
+## 0.2.0 -- 10/23/2016
 
 **Added**
 
@@ -10,7 +19,7 @@
     * Readme typos.
 
 
-## 0.1.0 -- 10/22/2026
+## 0.1.0 -- 10/22/2016
 
 **Added**